X DevAPI: add isOpen() method to Session. allow session to be closed 	during test and avoid failed tearDown()

+++ src/test/java/testsuite/mysqlx/devapi/CollectionTest.java
@@ -53,7 +53,7 @@ public class CollectionTest extends DevApiBaseTestCase {
 
     @After
     public void teardownCollectionTest() {
-        if (this.isSetForMySQLxTests) {
+        if (this.isSetForMySQLxTests && this.session.isOpen()) {
             try {
                 dropCollection(this.collectionName);
             } catch (Exception ex) {
